4-A Shibuya crossing (time limit exceeded)
In terms of graph theory, we need to find the maximal size of a clique in the
given graph. This version has an algorithm based on the Bron-Kerbosch algorithm
(which in its original form serves to find all maximal cliques in a graph). To
speed things up, one could use a better "general" algorithm to find maximal
cliques, or - probably better - use that the given graph is a "circle graph".
(But even with this extra information, the problem does not seem to becom
trivial.)